dofkhkjfds

+++ DAB-ASS2-del2/Models/KeyResponsible.cs
@@ -7,7 +7,7 @@ namespace DAB_ASS2.Models
         [Key]
         public int key_responsible_ID { get; set; }
         public string key_responsible_name { get; set; }
-        public int key_responsible_CPR { get; set; }
+        public long key_responsible_CPR { get; set; }
         public string key_responsible_address { get; set; }
         public long key_responsible_phone { get; set; }
         public int key_responsible_photo_ID_number { get; set; }
